Use mutexes and threads to improve ScriptBlock cpu usage
ScriptBlock now uses a std::timed_mutex to wait for the process to exit.
The mutex gets unlocked by a SIGCHLD handler, if it's not unlocked on
time then the process is killed and marked as timed out.

This helps with cpu usage as there is no spinning loop checking if the
process exited.